requirements.txt
langchain==1.2.13
langgraph==1.1.3
langchain-openrouter==0.2.1
langchain-community==0.4.1
langchain-mcp-adapters==0.2.2
mcp==1.26.0
python-dotenv==1.2.2
tavily-python==0.7.23
